GEN_AI_Tasks/Task 2/code/requirements.txt
requests==2.32.5
requests-toolbelt==1.0.0
beautifulsoup4==4.14.2
google-genai==1.66.0
feedparser==6.0.12
